#9 refactor: 보안 관련 코드 AuthSecurityService로 분리
- 뷰 모델의 역할이 과중하여 별도의 객체로 분리했습니다.
